Move ServiceNetworkSender implementation
Move the service processor's ServiceNetworkSender implementations under
the processor module.  This can now be found at
service::processor::sender.
